requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   -r requirements.in
    #   discord-py
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
charset-normalizer==3.3.2
    # via requests
chess==1.10.0
    # via -r requirements.in
click==8.1.7
    # via nltk
dataclasses-json==0.6.7
    # via langchain-community
dateparser==1.2.0
    # via youtube-comment-downloader
discord-py==2.4.0
    # via -r requirements.in
diskcache==5.6.3
    # via pyautogen
distro==1.9.0
    # via openai
docker==7.1.0
    # via pyautogen
flaml==2.2.0
    # via pyautogen
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via playwright
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   langsmith
    #   openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.4.2
    # via openai
joblib==1.4.2
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.14
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.12
    # via -r requirements.in
langchain-core==0.2.34
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.22
    # via -r requirements.in
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.104
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.22.0
    # via dataclasses-json
mmh3==4.1.0
    # via pinecone-text
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nltk==3.9.1
    # via pinecone-text
numpy==1.26.4
    # via
    #   flaml
    #   langchain
    #   langchain-community
    #   pinecone-text
    #   pyautogen
openai==1.42.0
    # via
    #   langchain-openai
    #   pyautogen
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   langchain-core
    #   marshmallow
    #   pyautogen
pillow==10.4.0
    # via -r requirements.in
pinecone-client==5.0.1
    # via -r requirements.in
pinecone-plugin-inference==1.0.3
    # via pinecone-client
pinecone-plugin-interface==0.0.7
    # via
    #   pinecone-client
    #   pinecone-plugin-inference
pinecone-text==0.9.0
    # via -r requirements.in
playwright==1.46.0
    # via
    #   -r requirements.in
    #   playwright-stealth
playwright-stealth==1.0.6
    # via -r requirements.in
psutil==6.0.0
    # via -r requirements.in
pyautogen==0.2.35
    # via -r requirements.in
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pyautogen
pydantic-core==2.20.1
    # via pydantic
pyee==11.1.0
    # via playwright
python-dateutil==2.9.0.post0
    # via dateparser
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pinecone-text
    #   pyautogen
pytube==15.0.0
    # via -r requirements.in
pytz==2024.1
    # via
    #   -r requirements.in
    #   dateparser
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via
    #   dateparser
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   docker
    #   langchain
    #   langchain-community
    #   langsmith
    #   pinecone-text
    #   tiktoken
    #   youtube-comment-downloader
    #   youtube-transcript-api
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.32
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
termcolor==2.4.0
    # via pyautogen
tiktoken==0.7.0
    # via
    #   langchain-openai
    #   pyautogen
tqdm==4.66.5
    # via
    #   nltk
    #   openai
    #   pinecone-client
types-requests==2.32.0.20240712
    # via pinecone-text
typing-extensions==4.12.2
    # via
    #   langchain-core
    #   openai
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   pyee
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzlocal==5.2
    # via dateparser
urllib3==2.2.2
    # via
    #   docker
    #   pinecone-client
    #   requests
    #   types-requests
wget==3.2
    # via pinecone-text
yarl==1.9.4
    # via aiohttp
youtube-comment-downloader==0.1.76
    # via -r requirements.in
youtube-transcript-api==0.6.2
    # via -r requirements.in
